Decoupling apparel coloring from GAA
This change should allow users to bypass Garam Advanced Apparel as a method for preserving the coloring of clothing.  Someone will need to go through and create masks for the apparel that should follow material colors, I lack the skill do to so.  This does not address the variable clothing system GAA has in any way.